+++ lib/models/page_result/page_result.g.dart
@@ -1,0 +1,31 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'page_result.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+PageResult<T> _$PageResultFromJson<T>(
+  Map<String, dynamic> json,
+  T Function(Object? json) fromJsonT,
+) =>
+    PageResult<T>(
+      pageNumber: (json['pageNumber'] as num?)?.toInt(),
+      pageSize: (json['pageSize'] as num?)?.toInt(),
+      totalItems: (json['totalItems'] as num?)?.toInt(),
+      totalPages: (json['totalPages'] as num?)?.toInt(),
+      items: (json['items'] as List<dynamic>?)?.map(fromJsonT).toList(),
+    );
+
+Map<String, dynamic> _$PageResultToJson<T>(
+  PageResult<T> instance,
+  Object? Function(T value) toJsonT,
+) =>
+    <String, dynamic>{
+      'pageNumber': instance.pageNumber,
+      'pageSize': instance.pageSize,
+      'totalItems': instance.totalItems,
+      'totalPages': instance.totalPages,
+      'items': instance.items?.map(toJsonT).toList(),
+    };