+++ lib/utils/utils.dart
@@ -1,5 +1,4 @@
 export 'package:stoyco_shared/utils/buy_event_button.dart';
-export 'package:stoyco_shared/utils/colors.dart';
 export 'package:stoyco_shared/utils/expandable_button.dart';
 export 'package:stoyco_shared/utils/hex_color.dart';
 export 'package:stoyco_shared/utils/launch_location.dart';